# Backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: BeeBa API
  version: 0.1.0
  description: Versioned API for the BeeBa Basis / VR content catalog.
servers:
  - url: http://localhost:8080
    description: Local backend
  - url: https://beeba.org
    description: Production domain
paths:
...
  /api/v1/admin/comments/{comment_id}/approve:
    post:
      summary: Approve a pending comment and publish it
      operationId: approveComment
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Comment made visible and content comment counter updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/ModeratedComment"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/comments/{comment_id}/hide:
    post:
      summary: Hide a visible or pending comment
      operationId: hideComment
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationReviewRequest"
      responses:
        "200":
          description: Comment hidden and content comment counter adjusted if needed.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/ModeratedComment"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/admin/reports/{report_id}/status:
    post:
      summary: Update report review status
      operationId: reviewReport
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportReviewRequest"
      responses:
        "200":
          description: Report review status updated and audited.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                properties:
                  data:
                    $ref: "#/components/schemas/ModeratedReport"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: beeba_access_token
      description: Browser session cookie. Unsafe methods authenticated by cookie also require `X-CSRF-Token` matching the `beeba_csrf_token` cookie.
  responses:
    Error:
      description: Consistent API error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    ModerationReviewRequest:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
...
    ModeratedComment:
      type: object
      required:
        - comment_id
        - content_id
        - status
        - comments_count
        - updated_at
      properties:
        comment_id:
          type: string
          format: uuid
        content_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - visible
            - hidden
            - deleted
            - pending_moderation
        comments_count:
          type: integer
          minimum: 0
        reason:
          type: string
        updated_at:
          type: string
          format: date-time
...
    ReportReviewRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - in_review
            - resolved
            - rejected
        reason:
          type: string
          maxLength: 500
    ModeratedReport:
      type: object
      required:
        - report_id
        - status
        - updated_at
      properties:
        report_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - open
            - in_review
            - resolved
            - rejected
        resolved_by:
          type:
            - string
            - "null"
          format: uuid
        resolved_at:
          type:
            - string
            - "null"
          format: date-time
        reason:
          type: string
        updated_at:
          type: string
          format: date-time